Add docs and re-organise modules
