+++ src/device/mod.rs
@@ -1,2 +1,4 @@
+//! Concrete implementation of Human Interface Devices
+pub mod consumer;
 pub mod keyboard;
 pub mod mouse;
