--- src/device/mouse/descriptor.rs
@@ -1,35 +1,0 @@
-/// Implements a Hid Mouse that conforms to the Boot specification. This aims
-/// to be compatible with BIOS and other reduced functionality USB hosts
-///
-/// This is defined in Appendix B.1 of Device Class Definition for Human
-/// Interface Devices (Hid) Version 1.11 -
-/// <https://www.usb.org/sites/default/files/hid1_11.pdf> - Appendix B.2, E.10
-#[rustfmt::skip]
-pub const HID_BOOT_MOUSE_REPORT_DESCRIPTOR: &[u8] = &[
-    0x05, 0x01, // Usage Page (Generic Desktop),
-    0x09, 0x02, // Usage (Mouse),
-    0xA1, 0x01, // Collection (Application),
-    0x09, 0x01, //   Usage (Pointer),
-    0xA1, 0x00, //   Collection (Physical),
-    0x95, 0x03, //     Report Count (3),
-    0x75, 0x01, //     Report Size (1),
-    0x05, 0x09, //     Usage Page (Buttons),
-    0x19, 0x01, //     Usage Minimum (1),
-    0x29, 0x03, //     Usage Maximum (3),
-    0x15, 0x00, //     Logical Minimum (0),
-    0x25, 0x01, //     Logical Maximum (1),
-    0x81, 0x02, //     Input (Data, Variable, Absolute),
-    0x95, 0x01, //     Report Count (1),
-    0x75, 0x05, //     Report Size (5),
-    0x81, 0x01, //     Input (Constant),
-    0x75, 0x08, //     Report Size (8),
-    0x95, 0x02, //     Report Count (2),
-    0x05, 0x01, //     Usage Page (Generic Desktop),
-    0x09, 0x30, //     Usage (X),
-    0x09, 0x31, //     Usage (Y),
-    0x15, 0x81, //     Logical Minimum (-127),
-    0x25, 0x7F, //     Logical Maximum (127),
-    0x81, 0x06, //     Input (Data, Variable, Relative),
-    0xC0, //   End Collection,
-    0xC0, // End Collection
-];
