+++ src/lib.rs
@@ -1,4 +1,6 @@
-//! USB Hid devices
+//! Batteries included USB Human Interface Device (HID) library
+//!
+//! Based on [`usb-device`](<https://crates.io/crates/usb-device>)
 #![no_std]
 
 //Allow the use of std in tests
